move updateFavorites to lib
