+++ lib/updateFavorites.js
